+++ tests/Composer/Test/Package/Loader/ArrayLoaderTest.php
@@ -17,6 +17,12 @@ use Composer\Package\Dumper\ArrayDumper;
 
 class ArrayLoaderTest extends \PHPUnit_Framework_TestCase
 {
+
+    /**
+     * @var ArrayLoader
+     */
+    private $loader;
+
     public function setUp()
     {
         $this->loader = new ArrayLoader(null, true);
@@ -206,5 +212,45 @@ class ArrayLoaderTest extends \PHPUnit_Framework_TestCase
 
         $package = $this->loader->load($config);
         $this->assertFalse($package->isAbandoned());
+    }
+
+    public function pluginApiVersions()
+    {
+        return array(
+            array('1.0'),
+            array('1.0.0'),
+            array('1.0.0.0'),
+            array('1'),
+            array('=1.0.0'),
+            array('==1.0'),
+            array('~1.0.0'),
+            array('*'),
+            array('3.0.*'),
+            array('@stable'),
+            array('1.0.0@stable'),
+            array('^5.1'),
+            array('>=1.0.0 <2.5'),
+            array('x'),
+            array('1.0.0-dev'),
+        );
+    }
+
+    /**
+     * @dataProvider pluginApiVersions
+     */
+    public function testPluginApiVersionAreKeptAsDeclared($apiVersion)
+    {
+        $links = $this->loader->parseLinks('Plugin', '9.9.9', '', array('composer-plugin-api' => $apiVersion));
+
+        $this->assertArrayHasKey('composer-plugin-api', $links);
+        $this->assertSame($apiVersion, $links['composer-plugin-api']->getConstraint()->getPrettyString());
+    }
+
+    public function testPluginApiVersionDoesSupportSelfVersion()
+    {
+        $links = $this->loader->parseLinks('Plugin', '6.6.6', '', array('composer-plugin-api' => 'self.version'));
+
+        $this->assertArrayHasKey('composer-plugin-api', $links);
+        $this->assertSame('6.6.6', $links['composer-plugin-api']->getConstraint()->getPrettyString());
     }
 }
